[refactor] added bip32 package to compat package
