+++ compat/bip32/derive.go
@@ -1,4 +1,4 @@
-package bip32
+package compat
 
 import (
 	"errors"
