+++ compat/bip32/derive_test.go
@@ -1,4 +1,4 @@
-package bip32
+package compat
 
 import (
 	"errors"
