+++ compat/bip32/extendedkey.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package bip32
+package compat
 
 // References:
 //   [BIP32]: BIP0032 - Hierarchical Deterministic Wallets
