+++ compat/bip32/hd_key_test.go
@@ -1,4 +1,4 @@
-package bip32
+package compat
 
 import (
 	"encoding/hex"
